bags: remove update faction event
This event was introduced as a hackfix to
provide proper repaints of the bags. There
was an issue that bags are not fully loaded
on PLAYER_ENTERING_WORLD, therefore the
later triggered UPDATE_FACTION was used.

As the bagframes have seen alot of rewrite
and cleanups, it should be save to remove this.