gui: entire rewrite
An entire rewrite of the gui focused on reducing boilerplate
code, runtime optimizations and adding usability features.

Moved to a dynamically stackable layout which is usable with
less code. A new search feature allows to lookup specific
config values and directly navigate to them.

The changes are incompatible to existing thirdparty addons.