chat: disable game function that closes chat windows
The 'GetChatWindowInfo' function returns shown as 'false' while the UIParent is hidden (Alt+Z).
Based on that return value, the 'FloatingChatFrame_Update' would remove every chat frame,
that isn't tabbed while the interface is hidden. With this overwrite we prevent the original
'FloatingChatFrame_Update' function from being called during that time to avoid this behaviour.