nameplates: add option for vertical offset
Due to the secure frame requirements in TBC, this option only exists in vanilla.

The feature is also experimental, as it removes the OnEnter/OnLeave events. There
is a ':Click' that can be called on the parent, but there is no ':Enter' or ':Leave'
that could be called to run the scripts. The parent also have no GetScript("OnEnter")
so there's no way to access the script that changes the color and shows the tooltip.