map: skip coords while map is in unknown state
This is required to be compatible to the latest version of aero[1]. While using the addon,
the `WorldMapButton` does not return a proper `:GetCenter()` for a short period of time. This
leads to `mx` and `my` being nil and breaking the calculation. Now the coordinates will just
be skipped as long as `mx` and `my` could not be determined via `:GetCenter()`.

[1] https://github.com/gashole/Aero